FEAT: 결제 승인 요청 기능 예외처리 추가
