+++ src/main/java/com/sparta/deliveryi/payment/application/service/PaymentApplication.java
@@ -1,10 +1,10 @@
 package com.sparta.deliveryi.payment.application.service;
 
 import com.sparta.deliveryi.payment.application.dto.PaymentConfirmCommand;
-import com.sparta.deliveryi.payment.application.dto.PaymentConfirmResponse;
+import com.sparta.deliveryi.payment.application.dto.PaymentResponse;
 
 import java.util.UUID;
 
 public interface PaymentApplication {
-    PaymentConfirmResponse confirm(UUID userId, PaymentConfirmCommand command);
+    PaymentResponse confirm(UUID userId, PaymentConfirmCommand command);
 }
