+++ src/main/java/com/sparta/deliveryi/payment/application/service/TossPaymentsService.java
@@ -1,7 +1,7 @@
 package com.sparta.deliveryi.payment.application.service;
 
 import com.sparta.deliveryi.payment.application.dto.PaymentCancelRequest;
-import com.sparta.deliveryi.payment.infrastructure.dto.PaymentResponse;
+import com.sparta.deliveryi.payment.application.dto.PaymentResponse;
 
 public interface TossPaymentsService {
     PaymentResponse confirm(String paymentKey, String orderId, int amount);
