+++ src/main/java/com/sparta/deliveryi/payment/application/service/TossQueryService.java
@@ -1,6 +1,6 @@
 package com.sparta.deliveryi.payment.application.service;
 
-import com.sparta.deliveryi.payment.infrastructure.dto.PaymentResponse;
+import com.sparta.deliveryi.payment.application.dto.PaymentResponse;
 
 public interface TossQueryService {
     PaymentResponse getPaymentByOrderId(String orderId);
