+++ src/main/java/com/sparta/deliveryi/payment/infrastructure/api/TossPaymentsProperties.java
@@ -1,7 +1,7 @@
 package com.sparta.deliveryi.payment.infrastructure.api;
 
 import com.sparta.deliveryi.payment.application.dto.PaymentCancelRequest;
-import com.sparta.deliveryi.payment.infrastructure.dto.PaymentResponse;
+import com.sparta.deliveryi.payment.application.dto.PaymentResponse;
 import com.sparta.deliveryi.payment.application.service.TossPaymentsService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
