--- src/main/java/com/sparta/deliveryi/payment/infrastructure/dto/PaymentResponse.java
@@ -1,8 +1,0 @@
-package com.sparta.deliveryi.payment.infrastructure.dto;
-
-public record PaymentResponse (
-    int httpStatus,
-    String code,
-    String message,
-    PaymentInfo payment
-){}
